+++ src/ocean/text/util/StringSearch.d
@@ -789,33 +789,6 @@ struct StringSearch ( bool wide_char = false )
         return split_!(Char)(slices, str, delim, &locateChar, n, collapse);
     }
 
-    /**************************************************************************
-
-        ditto
-
-        Deprecated because it creates a new destination array of slices instead
-        of taking an existing one.
-
-        Params:
-             str      = input string
-             delim    = delimiter character
-             n        = maximum number of slices; set to 0 to indicate no limit
-             collapse = set to true to collapse consecutive occurrences to
-                        prevent producing empty "slices"
-
-        Returns:
-             the resulting slices
-
-     **************************************************************************/
-
-    deprecated Char[][] split ( Char[] str, Char delim, uint n = 0, bool collapse = false )
-    {
-        Char[][] slices;
-
-        split_!(Char)(slices, str, delim, &locateChar, n, collapse);
-
-        return slices;
-    }
 
     /**************************************************************************
 
@@ -838,27 +811,6 @@ struct StringSearch ( bool wide_char = false )
         return split(slices,  str, delim, n, true);
     }
 
-    /**************************************************************************
-
-        ditto
-
-        Deprecated because it creates a new destination array of slices instead
-        of taking an existing one.
-
-        Params:
-             str      = input string
-             delim    = delimiter character
-             n        = maximum number of slices; set to 0 to indicate no limit
-
-        Returns:
-             the resulting slices
-
-     **************************************************************************/
-
-    deprecated Char[][] splitCollapse ( Char[] str, Char delim, uint n = 0 )
-    {
-        return split(str, delim, n, true);
-    }
 
     /**************************************************************************
 
@@ -881,33 +833,6 @@ struct StringSearch ( bool wide_char = false )
         return split_!(Char[])(slices, str, delims, &locateCharSet, n, collapse);
     }
 
-    /**************************************************************************
-
-        ditto
-
-        Deprecated because it creates a new destination array of slices instead
-        of taking an existing one.
-
-        Params:
-             str      = input string
-             delim    = delimiter character
-             n        = maximum number of slices; set to 0 to indicate no limit
-             collapse = set to true to collapse consecutive occurrences to
-                        prevent producing empty "slices"
-
-        Returns:
-             the resulting slices
-
-     **************************************************************************/
-
-    deprecated Char[][] split ( Char[] str, Char[] delims, uint n = 0, bool collapse = false )
-    {
-        Char[][] slices;
-
-        split_!(Char[])(slices, str, delims, &locateCharSet, n, collapse);
-
-        return slices.dup;
-    }
 
     /**************************************************************************
 
@@ -931,27 +856,6 @@ struct StringSearch ( bool wide_char = false )
         return split(slices, str, delim, n, true);
     }
 
-    /**************************************************************************
-
-        ditto
-
-        Deprecated because it creates a new destination array of slices instead
-        of taking an existing one.
-
-        Params:
-             str      = input string
-             delim    = delimiter character
-             n        = maximum number of slices; set to 0 to indicate no limit
-
-        Returns:
-             the resulting slices
-
-     **************************************************************************/
-
-    deprecated Char[][] splitCollapse ( Char[] str, Char[] delim, uint n = 0 )
-    {
-        return split(str, delim, n, true);
-    }
 
     /**************************************************************************